setup_for_fluent as a module function
